pool: use an immutable hash table to keep track of deadlines
Significantly improves performance when a pool has many consumers.

Thanks to @mflatt for helping me understand the reason for the
performance difference here.

A modification to a mutable hash table invalidates some internal
caching necessary for iteration. So, an interleaving of modifications
and iterations on a mutable hash table is a pathologically-slow case
for mutable hash tables in Racket.

https://gist.github.com/Bogdanp/f199a11fe65e3b9a06f49da29c1dcae1
https://discord.com/channels/571040468092321801/893314076346826852/1399725507506864250